backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
spacy
transformers
torch
dateparser
pydantic
python-dotenv
python-multipart
groq>=0.4.0
google-genai>=1.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
